+++ changelog.md
@@ -1,3 +1,17 @@
+## 2020-09-03
+
+++ Version 1.1 "Socks with sandals" release
+
+++ Added terminal input for first and last names
+
+++ Updated README to include syntax for terminal arguments
+
+++ Added calculation for generations
+
+++ Added terminal commands for syntax help and version checking
+
+-- Removed check for + instead of - for 10 number social security number
+
 ## 2020-08-31
 
 ++ Established changelog